Only attempt to kill pods when they are running. Otherwise this might cause noisy kubectl error messages.

+++ scripts/k8s/kill-pod.sh
@@ -1,7 +1,11 @@
-#! /bin/bash
+#!/usr/bin/env bash
 
 NAME=$1
 NAMESPACE=${2:-stackrox}
 
 mapfile -t pods < <(kubectl -n "${NAMESPACE}" get po --selector app="${NAME}" -o jsonpath='{.items[].metadata.name}')
-kubectl -n "${NAMESPACE}" delete po "${pods[*]}" --grace-period=0
+if [[ ${#pods[@]} -gt 0 ]]; then
+    kubectl -n "${NAMESPACE}" delete po "${pods[*]}" --grace-period=0
+else
+    echo "No pods to terminate"
+fi
